[NET-512] Restructure framework generation
