+++ cmdb/framework/results/__init__.py
@@ -14,10 +14,8 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .type import TypeModel
-from .category import CategoryModel
+from .iteration import IterationResult
 
 __all__ = [
-    CategoryModel,
-    TypeModel
+    IterationResult
 ]
