+++ cmdb/interface/rest_api/framework_routes/type_routes.py
@@ -21,7 +21,7 @@ from flask import abort, request, current_app
 from cmdb.framework.models.type import TypeModel
 from cmdb.framework.managers import ManagerGetError, ManagerInsertError, ManagerUpdateError, ManagerDeleteError
 from cmdb.framework.managers.error.framework_errors import FrameworkIterationError
-from cmdb.framework.managers.results import IterationResult
+from cmdb.framework.results.iteration import IterationResult
 from cmdb.framework.managers.type_manager import TypeManager
 from cmdb.framework.utils import PublicID
 from cmdb.interface.api_parameters import CollectionParameters
